feat: implement delete address usecase

+++ src/application/deleteAddressFromSession.ts
@@ -1,17 +1,64 @@
-import type { Address } from "../domains/Address";
+import type { AddressID } from "../domains/Address";
 import { useDeleteAddress } from "../services/api";
+import { useTempEmail } from "../services/tempEmailAdaptor";
+import { useNotifier } from "../services/notificationAdaptor";
+import { AlertDialogService, NotificationService } from "./ports";
+import { useAlertDialog } from "../services/alertDialog";
 
 function useDeleteAddressFromSession() {
-  const [deleteAddress, { data: addressWithRestoreKey, error }] =
-    useDeleteAddress();
+  const notifier: NotificationService = useNotifier();
+  const dialog: AlertDialogService = useAlertDialog();
+  const { sessionID: sessionId } = useTempEmail();
+  const [deleteAddress, { data: apiData, loading: apiLoading, error, called }] =
+    useDeleteAddress(sessionId);
 
-  async function doDeleteAddress(address: Address) {
-    try {
-      await deleteAddress({
-        variables: { input: { addressId: address.id } },
-      });
-    } catch (e) {
-      throw new Error(error?.message);
+  // TODO: fix console error
+  // Uncaught (in promise) Error
+  // at doDeleteAddress (deleteAddressFromSession.ts:21:13)
+  //
+  // TODO: refactor this use-case
+  // TODO: add notification whether it passes or not
+  // TODO: Uncaught (in promise) Error: ApolloError: invalid_id
+  // TODO: Uncaught (in promise) TypeError: Cannot read properties
+  // of undefined (reading 'deleteAddress
+  //
+  // TODO: async loading status for delete button on dialog ✅
+  // TODO: code-smell
+  // TODO: implement abort-controller with timeout for dialog
+  async function doDeleteAddress(address: AddressID) {
+    const action = await dialog.openDialog({
+      body: "Are you sure?",
+      header: "Delete Address",
+    });
+
+    if (action) {
+      dialog.isLoading();
+
+      try {
+        let response = await deleteAddress({
+          variables: { input: { addressId: address.id } },
+        });
+
+        const isDeleted = response.data.deleteAddress;
+
+        if (isDeleted) {
+          console.log("here in if");
+          notifier.notifySuccess("Done.", "Address succesfully removed.");
+          dialog.closeDialog();
+        } else {
+          dialog.closeDialog();
+          notifier.notifyError(
+            "Failed.",
+            "Couldn't remove the address, try later!"
+          );
+        }
+      } catch (e: any) {
+        throw new Error(e, {});
+      }
+    } else {
+      // Decline silently
+      dialog.closeDialog();
+      console.log("declined.");
     }
   }
